Added mute button to the mic
Wanted to have a mute button for my mic that just switched the volume from max to min. I added another row in the analogSliderValues array for the mic. This additional row is set to either 0 or 1024 if the mute button is toggled. There is also an led that indicates that the mute button is toggled for on/mute.